Small changes to src code now that im running Linux + Taking a break from VoxEng stuff

+++ Src/Engine/RenderFrame.cc
@@ -62,7 +62,7 @@ void Engine::render_frame()
     }
 
     //Update moving lights
-    for (int i = 0; i < lights.size(); i++)
+    for (size_t i = 0; i < lights.size(); i++)
     {
         if (lights[i].follower != nullptr)
         {
